add new module web. with spring Boot application.
