+++ .idea/compiler.xml
@@ -9,6 +9,7 @@
         <outputRelativeToContentRoot value="true" />
         <module name="core" />
         <module name="console" />
+        <module name="web" />
       </profile>
     </annotationProcessing>
   </component>
@@ -16,6 +17,7 @@
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
       <module name="console" options="-parameters" />
       <module name="core" options="-parameters" />
+      <module name="web" options="-parameters" />
     </option>
   </component>
 </project>
