--- core/src/main/java/acadamy/learnprogramming/MessageGenerator.java
@@ -1,10 +1,0 @@
-package acadamy.learnprogramming;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public interface MessageGenerator {
-    String getMainMessage();
-
-    String getResultMessage();
-}
