--- core/src/main/java/acadamy/learnprogramming/MinNumber.java
@@ -1,14 +1,0 @@
-package acadamy.learnprogramming;
-
-import org.springframework.beans.factory.annotation.Qualifier;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-@Qualifier
-public @interface MinNumber {
-}
