--- core/src/main/java/acadamy/learnprogramming/NumberGenerator.java
@@ -1,8 +1,0 @@
-package acadamy.learnprogramming;
-
-public interface NumberGenerator {
-
-    int next();
-    int getMaxNumber();
-    int getMinNumber();
-}
