+++ core/src/main/java/academy/learnprogramming/GameImpl.java
@@ -1,0 +1,86 @@
+package academy.learnprogramming;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+
+@Slf4j
+@Getter
+@Component
+public class GameImpl implements Game {
+
+    //    == fields
+    @Getter(AccessLevel.NONE)
+    private final NumberGenerator numberGenerator;
+    private final int guessCount;
+    private int number;
+    private int smallest;
+    private int biggest;
+    private int remainingGuesses;
+    private boolean validNumberRange = true;
+    @Setter
+    private int guess;
+
+    //    constructor
+    @Autowired
+    public GameImpl(NumberGenerator numberGenerator, @GuessCount int guessCount) {
+        this.numberGenerator = numberGenerator;
+        this.guessCount = guessCount;
+    }
+
+    @Override
+    public int getRemainingGuess() {
+        return remainingGuesses;
+    }
+
+    //    == init ==
+    @PostConstruct
+    @Override
+    public void reset() {
+        smallest = numberGenerator.getMinNumber();
+        guess = numberGenerator.getMinNumber();
+        remainingGuesses = guessCount;
+        biggest = numberGenerator.getMaxNumber();
+        number = numberGenerator.next();
+        log.debug("the number is {}", number);
+    }
+
+    @PreDestroy
+    public void preDestroy() {
+        log.info("in game preDestroy()");
+    }
+
+    @Override
+    public void check() {
+        checkValidNumber();
+        if (validNumberRange) {
+            if (guess > number) {
+                biggest = guess - 1;
+            } else {
+                smallest = guess + 1;
+            }
+        }
+        remainingGuesses--;
+    }
+
+    @Override
+    public boolean isGameWon() {
+        return guess == number;
+    }
+
+    @Override
+    public boolean isGameLoss() {
+        return !isGameWon() && remainingGuesses <= 0;
+    }
+
+    //    == private methods ==
+    private void checkValidNumber() {
+        validNumberRange = (guess >= smallest) && (guess <= biggest);
+    }
+}
