+++ core/src/main/java/academy/learnprogramming/MessageGenerator.java
@@ -1,0 +1,10 @@
+package academy.learnprogramming;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public interface MessageGenerator {
+    String getMainMessage();
+
+    String getResultMessage();
+}
