+++ core/src/main/java/academy/learnprogramming/NumberGenerator.java
@@ -1,0 +1,8 @@
+package academy.learnprogramming;
+
+public interface NumberGenerator {
+
+    int next();
+    int getMaxNumber();
+    int getMinNumber();
+}
