+++ pom.xml
@@ -19,6 +19,7 @@
     <modules>
         <module>core</module>
         <module>console</module>
+        <module>web</module>
     </modules>
     <packaging>pom</packaging>
