+++ web/src/main/java/academy/learnprogramming/WebMain.java
@@ -1,0 +1,16 @@
+package academy.learnprogramming;
+
+
+import com.sun.tools.javac.Main;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebMain {
+
+    public static void main(String[] args) {
+
+        SpringApplication.run(WebMain.class,args);
+
+    }
+}
