+++ web/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+#enable ansi output
+spring.output.ansi.enabled=always
